Some small refactoring.
